weather_bot/requirements.txt
aiogram==3.20.0.post0
python-dotenv
psycopg2-binary
SQLAlchemy
colorlog
emoji
httpx
aiohttp
requests